Move capi to separte static lib
